Make policy attachments depend on users
I recently added a user, and when I ran terraform apply it failed
because aws_iam_user.users["new.user"] did not yet exist.  To fix
this, I made the policy attachments depend on the aws_iam_user.users
resource.  This forces all the users to be created before attempting
to attach policies to them.